+++ packages/service-jwt/src/hooks.js
@@ -1,0 +1,5 @@
+const SERVICE_NAME = `jwt`;
+
+module.exports = {
+  SERVICE_NAME,
+};