+++ packages/kitchensink/CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - Added examples for `@forrestjs/hooks`
 - Added examples for `@forrestjs/service-fastify`
+- Added examples for `@forrestjs/service-jwt`
 - Added `app01` examples for a complex app